Import original (circa 2018) implementation
This is a prototype-quality implementation of the concept originally
written in 2018.

There are many, many things that need to be updated. At the time it was
written async rust wasn't yet stable, though it would be massively
beneficial for this use case.

The entire "pack" file format needs a redesign, or to be replaced with a
standard archive format.

Unfortunately Cargo.lock wasn't checked in, and Cargo can't currently
build it due to locked crates.